Improve PromQL queries on Grafana group labels
There is a push_job name defined in the apis.yml file. This job is used
as a label to identify metrics on Prometheus. However, this job name is
flexible and can be changed by the admins. The issue appears when this
job name is changed when data was already collected. In this case,
Grafana, by default, considers two different metrics and consequently
represent them separated on charts. The PromQL queries were updated to
deal with cases like this.