Change last by *last in charts and cleaned no value field
Depending on the chosen timeframe selected, the "last" metrics brings no
value likely due to the interval the metrics are collected. The field
"last" was changed by "*last" to ensure the last non null value is
always shown. In pizza charts was also cleaned the "no data" field. It
was already undefined but something happened during the Grafana updgrade
and manual intervention was necessary.